+++ src/main/java/http/request/HttpUrl.java
@@ -1,6 +1,5 @@
 package http.request;
 
-import stringprocessor.Params;
 import utils.IOUtils;
 
 public class HttpUrl {
