+++ src/main/java/http/request/Params.java
@@ -1,4 +1,4 @@
-package stringprocessor;
+package http.request;
 
 import java.util.HashMap;
 import java.util.Map;
