+++ src/test/java/http/request/ParamsTest.java
@@ -1,4 +1,4 @@
-package stringprocessor;
+package http.request;
 
 import static org.assertj.core.api.Assertions.*;
